fix: add Dock prompt preview popover
