+++ Sources/DockPanelView.swift
@@ -861,70 +861,6 @@ private struct DockErrorView: View {
     }
 }
 
-private struct DockEmptyView: View {
-    var body: some View {
-        VStack(spacing: 10) {
-            Image(systemName: "dock.rectangle")
-                .font(.system(size: 24))
-                .foregroundStyle(.secondary)
-            Text(String(localized: "dock.empty.title", defaultValue: "No Dock Controls"))
-                .font(.system(size: 13, weight: .semibold))
-            Text(String(
-                localized: "dock.empty.subtitle",
-                defaultValue: "Add controls to .cmux/dock.json."
-            ))
-            .font(.system(size: 12))
-            .foregroundStyle(.secondary)
-            HStack(spacing: 8) {
-                Button {
-                    copyAgentPrompt()
-                } label: {
-                    Label(
-                        String(localized: "dock.empty.copyPrompt", defaultValue: "Copy Agent Prompt"),
-                        systemImage: "doc.on.doc"
-                    )
-                }
-                .buttonStyle(.bordered)
-                .controlSize(.small)
-                .help(String(localized: "dock.empty.copyPrompt.help", defaultValue: "Copy a prompt you can paste into an AI coding agent"))
-
-                Button {
-                    openDockDocs()
-                } label: {
-                    Label(
-                        String(localized: "dock.empty.openDocs", defaultValue: "Docs"),
-                        systemImage: "questionmark.circle"
-                    )
-                }
-                .buttonStyle(.bordered)
-                .controlSize(.small)
-                .help(String(localized: "dock.empty.openDocs.help", defaultValue: "Open the Dock documentation"))
-            }
-            .padding(.top, 4)
-        }
-        .padding(20)
-        .frame(maxWidth: .infinity, maxHeight: .infinity)
-    }
-
-    private func copyAgentPrompt() {
-        let pasteboard = NSPasteboard.general
-        pasteboard.clearContents()
-        pasteboard.setString(agentPrompt, forType: .string)
-    }
-
-    private func openDockDocs() {
-        guard let url = URL(string: "https://cmux.com/docs/dock") else { return }
-        NSWorkspace.shared.open(url)
-    }
-
-    private var agentPrompt: String {
-        String(
-            localized: "dock.empty.agentPrompt",
-            defaultValue: "Add cmux Dock controls for this repo. Create .cmux/dock.json with a controls array. Each control needs id, title, command, and optional cwd, height, env. Use safe commands for this repo and do not include secrets. Include a Feed control only if cmux feed tui --opentui is useful. Add useful controls for this project, such as git, logs, dev server status, task queue, or test watcher."
-        )
-    }
-}
-
 private struct DockKeyboardFocusBridge: NSViewRepresentable {
     @ObservedObject var store: DockControlsStore
